Refactor: Move client-side utilities to $lib/client directory
